bench: split encode/decode bench w/wo serde feature
